requirements.txt
pandas
yfinance
sqlalchemy
pyodbc
python-dotenv
pytz
requests